requirements.txt
gradio==6.13.0
torch==2.11.0
torchvision==0.26.0
numpy==2.4.4
pillow==12.2.0
grad-cam==1.5.5
opencv-python==4.13.0.92
reportlab==4.4.5
